% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_all.R
\name{get_data_all}
\alias{get_data_all}
\alias{get_all_events}
\alias{get_all_matches}
\alias{get_all_players}
\alias{get_all_player_matches}
\alias{get_all_pro_players}
\alias{get_all_ama_players}
\title{Get all data for unit type}
\usage{
get_all_events()

get_all_matches(sleep_sec = 1, progress = TRUE)

get_all_players()

get_all_player_matches(player_id = integer(0))

get_all_pro_players()

get_all_ama_players()
}
\arguments{
\item{sleep_sec}{Number of seconds to wait before the next query}

\item{progress}{Boolean indicator of whether to show function progress}

\item{player_id}{Player id in character, integer or numeric form}
}
\value{
\code{get_all_events} returns data.frame of \link[=event]{events}
  for all seasons.

  \code{get_all_matches} returns data.frame of \link[=match]{matches}
  for all seasons.

  \code{get_all_players} returns data.frame of \link[=player]{players}
  for all seasons. Column \code{status} is correctly filled with
  \code{pro} and \code{ama}.

  \code{get_all_player_matches} returns data.frame of \link[=match]{matches}
  for all seasons for a specific player.

  \code{get_all_pro_players} returns data.frame of professional
  \link[=player]{players} for all seasons. Column \code{status} is filled
  with \code{pro}.

  \code{get_all_ama_players} returns data.frame of amateur
  \link[=player]{players} for all seasons. Column \code{status} is filled
  with \code{ama}.
}
\description{
These functions return all available data for a specific unit.
}
\details{
Service api.snooker.org doesn't have an option of querying all
  matches with one request. \code{get_all_matches} performs sequential queries
  over all events for their matches and combines them afterwards.
  \code{sleep_sec} is introduced in order to not overload API servers.
  It can be not less than 1 second. If \code{progress == TRUE} text progress
  will be displayed in the console.

  \strong{Note}: \code{get_all_matches} takes a lot of time to run.
}
\examples{
\dontrun{
# Get events for all seasons
get_all_events()
}
}
\seealso{
\link{get_data_by_id} for querying data by ID.

\link{get_data_multiple} for querying multiple data for one unit.
}
