% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_by_id.R
\name{get_data_by_id}
\alias{get_data_by_id}
\alias{get_events}
\alias{get_matches}
\alias{get_players}
\title{Get data for a unit by ID}
\usage{
get_events(event_id = integer(0), sleep_sec = 1, progress = TRUE)

get_matches(event_id = integer(0), round_id = integer(0),
  match_num = integer(0), sleep_sec = 1, progress = TRUE)

get_players(player_id = integer(0), sleep_sec = 1, progress = TRUE)
}
\arguments{
\item{event_id}{Vector of event ids in character, integer or numeric form}

\item{sleep_sec}{Number of seconds to wait before the next query}

\item{progress}{Boolean indicator of whether to show function progress}

\item{round_id}{Vector of round ids in character, integer or numeric form}

\item{match_num}{Vector of match numbers in character, integer or
numeric form}

\item{player_id}{Vector of player ids in character, integer or numeric form}
}
\value{
\code{get_events} returns \link{event} data.frame. Rows are arranged
  from the latest ended event to the earliest.

  \code{get_matches} returns \link{match} data.frame. Rows are arranged
  from the latest ended match to the earliest.

  \code{get_players} returns \link{player} data.frame.
  Column \code{status} is filled with \code{NA} because
  there is no information from performed query.
  Information about player status can be obtained from calls to
  \link[=get_data_multiple]{get_season_pro_players},
  \link[=get_data_multiple]{get_season_ama_players},
  \link[=get_data_all]{get_all_players},
  \link[=get_data_all]{get_all_pro_players},
  \link[=get_data_all]{get_all_ama_players}.
}
\description{
These functions return data about events, matches or players by
the vector(s) of respective identifiers.
}
\details{
Identifier for \strong{event} is \code{event_id}.

  Identifier for \strong{match} is combination of \code{event_id},
  \code{round_id} and \code{match_num}. When using \code{get_matches} these
  vectors are recycled by standard R rules.

  Identifier for \strong{player} is \code{player_id}.

  If there is no available data for particular identifier then it
  will not be included in return value (with appropriate message).

  \code{sleep_sec} is introduced in order to not overload API servers.
  It can be not less than 1 second. If \code{progress == TRUE} text progress
  will be displayed in the console.
}
\examples{
\dontrun{
# Get data about events with IDs 1 and 2
get_events(event_id = 1:2)

# Get data about match in event 398, round 4 and match number 1
get_matches(event_id = 398, round_id = 4, match_num = 1)

# Get data about players with IDs from 1 to 10
get_players(player_id = 1:10)
}
}
\seealso{
\link{get_data_multiple} for querying multiple data for one unit.

\link{get_data_all} for querying all data for unit type.
}
