% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_multiple.R
\name{get_data_multiple}
\alias{get_data_multiple}
\alias{get_season_events}
\alias{get_event_matches}
\alias{get_season_player_matches}
\alias{get_event_players}
\alias{get_season_pro_players}
\alias{get_season_ama_players}
\title{Get multiple data for one unit}
\usage{
get_season_events(season_id)

get_event_matches(event_id)

get_season_player_matches(season_id, player_id)

get_event_players(event_id)

get_season_pro_players(season_id)

get_season_ama_players(season_id)
}
\arguments{
\item{season_id}{Year of a season start}

\item{event_id}{Event id in character, integer or numeric form}

\item{player_id}{Player id in character, integer or numeric form}
}
\value{
\code{get_season_events} returns data.frame of \link[=event]{events} in
  a specific season.

  \code{get_event_matches} returns data.frame of \link[=match]{matches} in
  a specific event.

  \code{get_season_player_matches} returns data.frame of \link[=match]{matches} for
  a specific player in a specific season.

  \code{get_event_players} returns data.frame of \link[=player]{players} in
  a specific event. Column \code{status} is filled with \code{NA}.

  \code{get_season_pro_players} returns data.frame of professional
  \link[=player]{players} in a specific season. Column \code{status} is
  filled with \code{pro}.

  \code{get_season_ama_players} returns data.frame of amateur
  \link[=player]{players} in a specific season. Column \code{status} is
  filled with \code{ama}.
}
\description{
These functions return multiple data for a specific unit.
}
\details{
Seasons in snooker start approximately in May-June and end
  not long after World Championship ends.
}
\examples{
\dontrun{
# Get events that were played in season 2015-2016.
get_season_events(season_id = 2015)

# Get matches that were played in event with id 398.
get_event_players(event_id = 398)
}
}
