% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_single.R
\name{get_data_single}
\alias{get_data_single}
\alias{get_event}
\alias{get_match}
\alias{get_player}
\title{Get data for a single unit}
\usage{
get_event(event_id)

get_match(event_id, round_id, match_num)

get_player(player_id)
}
\arguments{
\item{event_id}{Event id in character, integer or numeric form}

\item{round_id}{Round id in character, integer or numeric form}

\item{match_num}{Match number in character, integer or numeric form}

\item{player_id}{Player id in character, integer or numeric form}
}
\value{
\code{get_event} returns a data.frame with 1 row and columns
  described in \link{event}.

  \code{get_match} returns a data.frame with 1 row and columns
  described in \link{match}.

  \code{get_player} returns a data.frame with 1 row and columns
  described in \link{player}. Column \code{status} is filled with
  \code{NA} because there is no information from performed query.
  Information about player status can be obtained from calls to
  \link[=get_data_multiple]{get_season_pro_players},
  \link[=get_data_multiple]{get_season_ama_players},
  \link[=get_data_all]{get_all_players},
  \link[=get_data_all]{get_all_pro_players},
  \link[=get_data_all]{get_all_ama_players}.
}
\description{
These functions return data about event, match or player by
the respective identifiers.
}
\details{
Identifier for \strong{event} is \code{event_id}.

  Identifier for \strong{match} is combination of \code{event_id},
  \code{round_id} and \code{match_num}.

  Identifier for \strong{player} is \code{player_id}.
}
\examples{
\dontrun{
# Get data about event 398
get_event(event_id = 1)

# Get data about match in event 398, round 4 and match number 1
get_match(event_id = 398, round_id = 4, match_num = 1)

# Get data about player 1
get_player(player_id = 1)
}
}
\seealso{
\link{get_data_multiple} for querying multiple data for one unit.

\link{get_data_all} for querying all data for unit type.
}
