% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{query_snookerorg}
\alias{query_snookerorg}
\title{Query api.snooker.org}
\usage{
query_snookerorg(...)
}
\arguments{
\item{...}{Parameters for query in character form.}
}
\value{
Raw query result from api.snooker.org.
}
\description{
Query api.snooker.org
}
\examples{
# Returns NULL
queryJSON()

\dontrun{
# Returns query for event with id 398.
query_snookerorg("e=398")
}
}
\keyword{internal}
