% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\alias{simplify_event}
\alias{simplify_match}
\alias{simplify_player}
\title{Simplify results of get_* functions}
\usage{
simplify_event(event_data)

simplify_match(match_data)

simplify_player(player_data)
}
\arguments{
\item{event_data}{data.frame of type \link{event}}

\item{match_data}{data.frame of type \link{match}}

\item{player_data}{data.frame of type \link{player}}
}
\value{
Simplified versions of specific data.frames with selecting
  the mostly used columns.
}
\description{
Simplify results of get_* functions
}
\examples{
\dontrun{
simplify_event(get_event(event_id = 398))
}

}
